Improve make_optional_func_list and make renegotiation functions optional.
`make_optional_func_list` will fall back to using `nm` to determine if a
function is provided by libwolfssl if libwolfssl is a static library. This
commit makes it so we first check if `nm` is available. If it's not available,
the function will print a warning and assume all optional functions are
available.